requirements.txt
#requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
celery[redis]
redis
